refactor: ♻️ extract shared modal logic into FoodCardShell, add dietary conflicts to PopularDishCard
